Add deleting image functions

+++ backend/app/routes/route.py
@@ -751,27 +751,28 @@ class FileView(MethodView):
                 if not os.path.isdir(images):
                     os.mkdir(images)
                 image_path = os.path.join(images, 'image.jpg')
-                if image_path:
+                if os.path.isfile(image_path):
                     os.remove(image_path)
-                rgb_im.save(os.path.join(image_path))
+                rgb_im.save(image_path)
             else:
                 for file in files:
                     filename = secure_filename(file.filename)
                     for file in files:
                         file.save(os.path.join(folder, 
                             f'{datetime.now().strftime("%Y-%m-%d %H-%M-%S")}-{filename}'))
-            return '', 201
+            return {'message': item_id}
         
 
     @r_g.group_required(Groups.staffsec.name)
     @r_g.roles_required(Roles.user.name)
     @bp.output(EmptySchema, status_code=204)
     def delete(self, action, item_id):
-        person = db.session.query(Person).get(item_id)
-        os.remove(person.path)
-        person.path = ''
-        db.session.commit()
-        return ''
+        if action == 'image':
+            person = db.session.query(Person).get(item_id)
+            os.remove(person.path)
+            person.path = ''
+            db.session.commit()
+            return ''
 
 bp.add_url_rule('/file/<action>/<int:item_id>', 
                 view_func=FileView.as_view('file'))
